feat: badge modify and display change
